backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
langchain==0.3.21
langchain-community==0.3.20
pymilvus==2.5.6
sentence-transformers==3.4.1
python-dotenv==1.0.1
huggingface-hub>=0.19.0
pymongo==4.6.2
passlib[bcrypt]==1.7.4